+++ packages/showcases/react/stories/components/structure/VtmnListItem/VtmnListItem.stories.tsx
@@ -1,0 +1,60 @@
+import React from 'react';
+import { Meta, Story } from '@storybook/react';
+import { parameters } from '@vtmn/showcase-core/csf/components/structure/list.csf';
+import { VtmnList, VtmnListItem, VtmnListItemText } from '@vtmn/react';
+
+export default {
+  title: 'Components / Structure / VtmnListItem',
+  component: VtmnListItem,
+  argTypes: {
+    divider: {
+      type: { name: 'boolean', required: false },
+      description: 'Whether to show the divider under the item',
+      defaultValue: false,
+      control: {
+        type: 'boolean',
+      },
+    },
+    disabled: {
+      type: { name: 'boolean', required: false },
+      description: 'The disabled state of the item',
+      defaultValue: false,
+      control: {
+        type: 'boolean',
+      },
+    },
+    size: {
+      type: { name: 'string', required: false },
+      description: 'Size of the item.',
+      defaultValue: 'medium',
+      control: {
+        type: 'select',
+        options: ['small', 'medium'],
+      },
+    },
+    children: {
+      type: { name: 'string', required: false },
+      description: 'The content to render inside the component.',
+      defaultValue: 'Label',
+      control: {
+        type: 'text',
+      },
+    },
+  },
+  parameters,
+} as Meta;
+
+const OverviewTemplate: Story = (args) => (
+  <div style={{ width: '500px' }}>
+    <VtmnList>
+      <VtmnListItem {...args}>
+        <VtmnListItemText onClick={() => console.log(`List item clicked`)}>
+          <span>Lorem ipsum</span>
+          <span>Lorem ipsum dolor sit amet</span>
+        </VtmnListItemText>
+      </VtmnListItem>
+    </VtmnList>
+  </div>
+);
+
+export const Overview = OverviewTemplate.bind({});