+++ packages/sources/svelte/src/components/overlays/VtmnModal/test/VtmnModal.spec.js
@@ -1,10 +1,8 @@
 import '@testing-library/jest-dom';
-
 import { fireEvent, render } from '@testing-library/svelte';
-
 import VtmnModal from '../VtmnModal.svelte';
-import VtmnModalWithoutActions from './VtmnModalWithoutActions.svelte';
 import VtmnModalWithActions from './VtmnModalWithActions.svelte';
+import VtmnModalWithoutActions from './VtmnModalWithoutActions.svelte';
 
 const expectedCloseOnElement = async (
   element,
@@ -24,10 +22,6 @@ describe('VtmnModal', () => {
     container.getElementsByClassName('vtmn-modal_background-overlay')[0];
   const getCloseButton = (container) =>
     container.querySelector('button[aria-label="close"]');
-  const getOverFlow = (container) =>
-    container.getElementsByClassName(
-      'vtmn-modal_content_body--overflow-indicator',
-    )[0];
   const getModalContent = (container) =>
     container.getElementsByClassName('vtmn-modal_content')[0];
   const getContentActions = (container) =>
@@ -123,21 +117,9 @@ describe('VtmnModal', () => {
         });
         expect(getContentActions(container)).toBeUndefined();
       });
-      test("Should not have a class 'vtmn-modal_content_body--overflow-indicator'", () => {
-        const { container } = render(VtmnModalWithoutActions, {
-          open: true,
-        });
-        expect(getOverFlow(container)).toBeUndefined();
-      });
     });
 
     describe('With slot description and actions', () => {
-      test("Should have a class 'vtmn-modal_content_body--overflow-indicator'", () => {
-        const { container } = render(VtmnModalWithActions, {
-          open: true,
-        });
-        expect(getOverFlow(container)).toBeVisible();
-      });
       test('Should have a slot actions', () => {
         const { container } = render(VtmnModalWithActions, {
           open: true,