+++ packages/sources/css/src/components/actions/dropdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-dropdown@0.4.1...@vtmn/css-dropdown@0.4.2) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-dropdown
+
+
+
+
+
 ## [0.4.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-dropdown@0.4.0...@vtmn/css-dropdown@0.4.1) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-dropdown