+++ packages/sources/css/src/components/indicators/badge/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.13...@vtmn/css-badge@0.10.14) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-badge
+
+
+
+
+
 ## [0.10.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-badge@0.10.12...@vtmn/css-badge@0.10.13) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-badge