+++ packages/sources/css/src/components/indicators/progressbar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-progressbar@0.4.13...@vtmn/css-progressbar@0.4.14) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-progressbar
+
+
+
+
+
 ## [0.4.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-progressbar@0.4.12...@vtmn/css-progressbar@0.4.13) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-progressbar