+++ packages/sources/css/src/components/indicators/rating/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.15](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.14...@vtmn/css-rating@0.3.15) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-rating
+
+
+
+
+
 ## [0.3.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-rating@0.3.13...@vtmn/css-rating@0.3.14) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-rating