+++ packages/sources/css/src/components/navigation/navbar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.21](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.20...@vtmn/css-navbar@0.1.21) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
 ## [0.1.20](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.19...@vtmn/css-navbar@0.1.20) (2022-04-21)
 
 **Note:** Version bump only for package @vtmn/css-navbar